Move Corruption's Layout Editor to Corruption
