+++ randovania/games/prime3/gui/corruption_help_widget.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from randovania.games.game import RandovaniaGame
 from randovania.games.prime3.gui.generated.games_tab_corruption_widget_ui import Ui_CorruptionGameTabWidget
-from randovania.gui.corruption_layout_editor import CorruptionLayoutEditor
+from randovania.games.prime3.gui.corruption_layout_editor import CorruptionLayoutEditor
 from randovania.gui.widgets.base_game_tab_widget import BaseGameTabWidget
